[Cache] Add generic TagAwareAdapter wrapper (replaces TagAwareRedisAdapter)
